Add a specimen with custom XML namespaces
This structure-specific data message from ESTAT has a "xmlns:u" that
gives the SDMX URN (not a URL!) of the DSD used in the message. This is
followed by tags like <u:DataSet> (as opposed to <mes:DataSet>). This
XML causes a parse error noticed and fixed in khaeru/sdmx#107.